Relative Wall Angle Detection
Currently testing viability of this feature which allows for 2 parallel ultrasonic sensors to calculate the rough estimated angle of the surface perpendicular to said sensors.

## CAUTION ##
There is a limitation to the angle that can be interpreted. Any surface interpreted as beyond roughly 45 degrees  is not always a valid reading. Ensure that robot always stays in a position that these angles cannot be achieved nor measured. The exact angle and necessary error checking are in the works.